make this compile on i386-linux with _LARGEFILE64_SOURCE for lseek64 ... Ioctl converted back to int; must be fixed for APPLE ...

+++ de.ibapl.jnhw.posix/src/main/native/IoctlFunctions.c
@@ -34,11 +34,11 @@ extern "C" {
     /*
      * Class:     de_ibapl_jnhw_unix_sys_Ioctl
      * Method:    ioctl
-     * Signature: (IJ)I
+     * Signature: (II)I
      */
     JNIEXPORT jint JNICALL Java_de_ibapl_jnhw_unix_sys_Ioctl_ioctl__IJ
-    (JNIEnv *env, __attribute__ ((unused)) jclass clazz, jint fd, jlong request) {
-        int result = ioctl(fd, (uint64_t)request);
+    (JNIEnv *env, __attribute__ ((unused)) jclass clazz, jint fd, jint request) {
+        int result = ioctl(fd, (uint32_t)request);
         if (result < 0) {
             throw_NativeErrorException(env, errno);
         }
@@ -48,14 +48,14 @@ extern "C" {
     /*
      * Class:     de_ibapl_jnhw_unix_sys_Ioctl
      * Method:    ioctl
-     * Signature: (IJLde/ibapl/jnhw/IntRef;)I
+     * Signature: (IILde/ibapl/jnhw/IntRef;)I
      */
     JNIEXPORT jint JNICALL Java_de_ibapl_jnhw_unix_sys_Ioctl_ioctl__IJLde_ibapl_jnhw_IntRef_2
-    (JNIEnv *env, __attribute__ ((unused)) jclass clazz, jint fd, jlong request, jobject intRef) {
+    (JNIEnv *env, __attribute__ ((unused)) jclass clazz, jint fd, jint request, jobject intRef) {
 
         int _intRef = GET_INT_REF_VALUE(intRef);
 
-        int result = ioctl(fd, (uint64_t)request, &_intRef);
+        int result = ioctl(fd, (uint32_t)request, &_intRef);
         SET_INT_REF_VALUE(intRef, _intRef);
         if (result < 0) {
             throw_NativeErrorException(env, errno);
